docker/requirements.txt
python-dateutil
slack-sdk
aiohttp
google-cloud-storage